# src/main/java/com/att/tdp/issueflow/project/ProjectController.java
package com.att.tdp.issueflow.project;

import com.att.tdp.issueflow.project.dto.ProjectCreateRequest;
import com.att.tdp.issueflow.project.dto.ProjectResponse;
import com.att.tdp.issueflow.project.dto.ProjectUpdateRequest;
import com.att.tdp.issueflow.project.dto.WorkloadEntry;
import jakarta.validation.Valid;
import java.util.List;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/** REST endpoints for the Project aggregate. ADMIN gates restore + list-deleted. */
@RestController
@RequestMapping("/projects")
public class ProjectController {

    private final ProjectService projectService;

    public ProjectController(ProjectService projectService) {
        this.projectService = projectService;
    }

    @GetMapping
    public List<ProjectResponse> list() {
        return projectService.findAll();
    }

    @PostMapping
    public ProjectResponse create(@Valid @RequestBody ProjectCreateRequest body) {
        return projectService.create(body);
    }

    @GetMapping("/{projectId}")
    public ProjectResponse get(@PathVariable Long projectId) {
        return projectService.findById(projectId);
    }

    @PatchMapping("/{projectId}")
    public ProjectResponse update(@PathVariable Long projectId,
                                  @Valid @RequestBody ProjectUpdateRequest body) {
        return projectService.update(projectId, body);
    }

    @DeleteMapping("/{projectId}")
    public void delete(@PathVariable Long projectId) {
        projectService.softDelete(projectId);
    }

    @GetMapping("/deleted")
    @PreAuthorize("hasRole('ADMIN')")
    public List<ProjectResponse> listDeleted() {
        return projectService.findDeleted();
    }

    @PostMapping("/{projectId}/restore")
    @PreAuthorize("hasRole('ADMIN')")
    public void restore(@PathVariable Long projectId) {
        projectService.restore(projectId);
    }

    @GetMapping("/{projectId}/workload")
    public List<WorkloadEntry> workload(@PathVariable Long projectId) {
        return projectService.workload(projectId);
    }
}